Implement variant derivation strategy
This commit changes how Maven metadata is derived into variants. Now we will
only derive variants if the Java plugin is applied (the "base Java" plugin).
This is implemented via a variant derivation strategy, and allows fixing
the problem that a native component is unlikely to find sense in the derivation
of Java variants. This fixes a bug where the native plugins wouldn't be able
to consume native libraries published on Maven repositories, without Gradle
metadata.